backends/native/kms-crtc: Don't compare gamma values on unitialized memory
On hotplugs, to check if we had a gamma change we compare the gamma
states values, and we assumed that the gamma pointers were initialized
by KMS, however drmModeCrtcGetGamma may not set those values, leaving
the gamma values not initialized.

And this may lead meta_kms_crtc_state_changes() to return wrong values.